# Optimize StatefulWriter.check_acked_status [4897]