Do a sequence number range check on each change in history rather than an == comparison.

Ran a comparison test, using vtune, on ros1bridge implementation with and without optimized code on a low-latency environment.

With Optimization:
Elapsed time: 386.000s
CPU Time: 459.260s
CPU Time in optimized path: 3.328s

Without Optimization:
Elapsed Time: 403.621s
CPU Time: 591.480
CPU Time in optimized path: 302.246