@araccp 2019-03-15
I have made the changes advised but these tests stream appear to be giving rather inconsistent results.  Technically, all three tests have passed but never all at once.  After some time of investigating, it seems to me that I am either unlucky or have exposed bugs; I am believe it's both.  I would really appreciate any additional insight or information regarding these failures that I'm getting.

The following two are not a concern but could be improved in the test code:
1) Linux has one failure - BlackBox_PreallocMem.AsyncPubSubAsReliableData300kb.  My investigation has found that there is a small, but existent, chance to fail when the fastrtps stack works as expected.  Ensuring that the correct number of packets are dropped for the test would take a bit more complexity in test_UDPv4Transport::packet_should_drop; otherwise, adjusting the ASSERT to accept a range instead of the set number is another way to reduce flakiness.
2) Windows BlackBox_PreallocMem.TCP_TLS is failing due to what appears to be an intermittent error when calling asio::ip::tcp::socket::send.  The errno is not printed in RTCPMessageManager::sendMessage() which would be nice. I'm suggesting this is intermittent because it seems to happen on another test I have seen in other windows tests (1322).  The windows tests have since passed, only furthering my assumption.

SEGFAULTS:
1) Windows has a segfault which occurred one 2 tests: BlackBox_ReallocMem.TCPMaxInitialPeer_P0_5_P4.  This is more concerning because there is a semblance of consistency with this failure and the result is a crash.  Looking at the logs, it looks like it's crashing on cleanup as the crash occurs after the main bulk of work is completed. (after discovery) I need a backtrace for the core dump in order to know more.
2) Mac tests, 3 out of 4 result in being "unstable".  All 3 are failing on the same test with different conditions: BlackBox_PreallocMem.TCP_TLS, BlackBox_DynMem.TCP_TLS, BlackBox_ReallocMem.TCP_TLS.  Looks like all these tests fail at the end (upon completion) leading me to suggest that there is an issue with cleanup.  Stacktraces are valuable for looking deeper into these segfaults.