@richiware 2019-03-18
Those errors are not related to your changes. We've updated our TCP transport with TLS support and it is a bit unstable. We are working on it. As long as our CI is stable, we will merge your changes. Thanks.