@MiguelCompany 2019-05-03
@araccp We are reducing the number of temporary allocations, and this PR includes the addition of a std::vector.

I have just created PR #515, which should improve performance. Would you mind to test that PR on the same environment specified on the description of this PR?