@araccp 2019-05-06
@MiguelCompany I'll have some time today to run a comparison.