@MiguelCompany 2019-05-21
Closing this, as #515 is now part of release v1.8.0. 

@araccp Feel free to check the timings with the new release and comment here.